requirements.txt
Django==4.0.3
unittest2==1.1.0
psycopg2==2.9.2
djangorestframework==3.13.1
factory-boy==3.2.1